Moved the MiniMapRenderer from Tiled to libtiled
