+++ src/libtiled/libtiled-src.pri
@@ -20,6 +20,7 @@ SOURCES += $$PWD/compression.cpp \
     $$PWD/maprenderer.cpp \
     $$PWD/maptovariantconverter.cpp \
     $$PWD/mapwriter.cpp \
+    $$PWD/minimaprenderer.cpp \
     $$PWD/object.cpp \
     $$PWD/objectgroup.cpp \
     $$PWD/objecttemplate.cpp \
@@ -63,6 +64,7 @@ HEADERS += $$PWD/compression.h \
     $$PWD/maprenderer.h \
     $$PWD/maptovariantconverter.h \
     $$PWD/mapwriter.h \
+    $$PWD/minimaprenderer.h \
     $$PWD/object.h \
     $$PWD/objectgroup.h \
     $$PWD/objecttemplate.h \
