+++ src/libtiled/minimaprenderer.cpp
@@ -2,22 +2,30 @@
  * minimaprenderer.cpp
  * Copyright 2017, Yuriy Natarov <natarur@gmail.com>
  * Copyright 2012, Christoph Schnackenberg <bluechs@gmx.de>
- * Copyright 2012, Thorbjørn Lindeijer <thorbjorn@lindeijer.nl>
+ * Copyright 2012-2020, Thorbjørn Lindeijer <thorbjorn@lindeijer.nl>
  *
- * This file is part of Tiled.
+ * This file is part of libtiled.
  *
- * This program is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License as published by the Free
- * Software Foundation; either version 2 of the License, or (at your option)
- * any later version.
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
  *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
- * more details.
+ *    1. Redistributions of source code must retain the above copyright notice,
+ *       this list of conditions and the following disclaimer.
  *
- * You should have received a copy of the GNU General Public License along with
- * this program. If not, see <http://www.gnu.org/licenses/>.
+ *    2. Redistributions in binary form must reproduce the above copyright
+ *       notice, this list of conditions and the following disclaimer in the
+ *       documentation and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE CONTRIBUTORS ``AS IS'' AND ANY EXPRESS OR
+ * IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
+ * MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO
+ * EVENT SHALL THE CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+ * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+ * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+ * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
+ * OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
+ * ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
 #include "minimaprenderer.h"
@@ -25,11 +33,11 @@
 #include "hexagonalrenderer.h"
 #include "imagelayer.h"
 #include "isometricrenderer.h"
+#include "map.h"
 #include "mapobject.h"
 #include "maprenderer.h"
 #include "objectgroup.h"
 #include "orthogonalrenderer.h"
-#include "preferences.h"
 #include "staggeredrenderer.h"
 #include "tilelayer.h"
 
@@ -39,7 +47,7 @@
 
 using namespace Tiled;
 
-MiniMapRenderer::MiniMapRenderer(Map *map)
+MiniMapRenderer::MiniMapRenderer(const Map *map)
     : mMap(map)
 {
     switch (map->orientation()) {
@@ -64,6 +72,11 @@ MiniMapRenderer::MiniMapRenderer(Map *map)
 MiniMapRenderer::~MiniMapRenderer()
 {
     delete mRenderer;
+}
+
+QSize MiniMapRenderer::mapSize() const
+{
+    return mRenderer->mapBoundingRect().size();
 }
 
 QImage MiniMapRenderer::render(QSize size, RenderFlags renderFlags) const
@@ -241,8 +254,6 @@ void MiniMapRenderer::renderToImage(QImage& image, RenderFlags renderFlags) cons
         painter.translate(-offset);
     }
 
-    if (drawTileGrid) {
-        Preferences *prefs = Preferences::instance();
-        mRenderer->drawGrid(&painter, mapBoundingRect, prefs->gridColor());
-    }
+    if (drawTileGrid)
+        mRenderer->drawGrid(&painter, mapBoundingRect, mGridColor);
 }
