+++ src/tiled/tiled.qbs
@@ -304,8 +304,6 @@ QtGuiApplication {
         "minimapdock.cpp",
         "minimapdock.h",
         "minimap.h",
-        "minimaprenderer.cpp",
-        "minimaprenderer.h",
         "movelayer.cpp",
         "movelayer.h",
         "movemapobject.cpp",
