+++ src/tiled/tilestampmodel.cpp
@@ -93,7 +93,6 @@ bool TileStampModel::setData(const QModelIndex &index, const QVariant &value, in
                 emit stampRenamed(stamp);
                 emit stampChanged(stamp);
                 return true;
-                break;
             default:
                 break;
             }
